+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## 3.7.0
 
 ### Minor Changes
+
 - 1511 | Remove Complementary PUE field from Ecomind
 - 1548 | New design for Visualize my results
+- 1549 | [Digital service] Same design for the pop up in the visualization
 
 ## 3.6.1