+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - 1511 | Remove Complementary PUE field from Ecomind
 - 1548 | New design for Visualize my results
 - 1549 | [Digital service] Same design for the pop up in the visualization
+- 1597 | add comma for VM
 
 ## 3.6.1